improved code organization and readability
